Use recommended gem structure
RubyGems and Bundler know how to handle gems that extend other gems
as long as their naming convention is followed. If the name of the
gem is active_remote-cached, they will attempt to load the gem with:

    require 'active_remote/cached'

Reorganize the gem structure to follow the convention. Leave
active_remote-cached.rb in case someone is using an older version of
RubyGems or Bundler.
