+++ active_remote-cached.gemspec
@@ -1,7 +1,7 @@
 # -*- encoding: utf-8 -*-
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'active_remote-cached/version'
+require 'active_remote/cached/version'
 
 Gem::Specification.new do |gem|
   gem.name          = "active_remote-cached"
