+++ lib/active_remote/cached/cache.rb
