+++ lib/active_remote/cached/version.rb
