+++ spec/active_remote/cached_delete_methods_spec.rb
