+++ spec/active_remote/cached_exist_methods_spec.rb
