+++ spec/active_remote/cached_search_methods_spec.rb
